Вынес классы из файла
